Update anderson_credit.pl
This script allows the user to enter a credit card number for the command line, then splits the credit card number in groups of four digits with each group printed on a new line without dashes.  Valid credit card numbers are: 123456781234643, 1111-2222-4444-9999 or any combination of the two.  The magic behind the scenes of this script exploits "capture variables", a unique feature of Perl that searches for and then splits the regex pattern by sub strings.